:bug: fix asciinema ssr failed
